ChainCell implemented
